requirements.txt
# Core dependencies for Decentralized Localization System

# Numerical computation
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0

# Optimization (for MPS algorithm)
cvxpy>=1.3.0

# Configuration management
pyyaml>=6.0

# Network analysis
networkx>=2.8

# Distributed computing (for MPI-based algorithms)
mpi4py>=3.1.0

# Optional: Progress bars
# tqdm>=4.60.0